requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-extensions==3.2.1
django-redis==5.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
jsonfield==3.1.0
mergedeep==1.3.4
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.4.0
